+++ spin2/src/spin1.semantic.ts
@@ -100,7 +100,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
 
   private spin1log: any = undefined;
   // adjust following true/false to show specific parsing debug
-  private spin1DebugLogEnabled: boolean = false; // WARNING (REMOVE BEFORE FLIGHT)- change to 'false' - disable before commit
+  private spin1DebugLogEnabled: boolean = true; // WARNING (REMOVE BEFORE FLIGHT)- change to 'false' - disable before commit
   private showSpinCode: boolean = true;
   private showPreProc: boolean = true;
   private showCON: boolean = true;
@@ -391,6 +391,8 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
         currSingleLineBlockComment = new RememberedComment(eCommentType.singleLineComment, i, line);
         continue;
       } else if (sectionStatus.isSectionStart) {
+        // mark end of method, if we were in a method
+        this.semanticFindings.endPossibleMethod(i); // pass prior line number!
         currState = sectionStatus.inProgressStatus;
         this._logState("- scan ln:" + (i + 1) + " currState=[" + currState + "]");
         // ID the remainder of the line
@@ -490,6 +492,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
         this._getSpinObjectConstantMethodDeclaration(0, i + 1, line);
       }
     }
+    this.semanticFindings.endPossibleMethod(lines.length - 1); // report end if last line of file
     // --------------------         End of PRE-PARSE             --------------------
     this._logMessage("---> Actual SCAN");
 
@@ -1086,7 +1089,11 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
     const methodName = line.substr(startNameOffset, nameLength).trim();
     const nameType: string = isPrivate ? "private" : "public";
     this._logSPIN("  -- GLBL GetMethodDecl newName=[" + methodName + "](" + methodName.length + "), type=[" + methodType + "], currentOffset=[" + currentOffset + "]");
+
     this.currentMethodName = methodName; // notify of latest method name so we can track inLine PASM symbols
+    // mark start of method - we are learning span of lines this method covers
+    this.semanticFindings.startMethod(methodName, lineNbr);
+
     // remember this method name so we can annotate a call to it
     const refModifiers: string[] = isPrivate ? ["static"] : [];
     this.semanticFindings.setGlobalToken(methodName, new RememberedToken("method", refModifiers), lineNbr, this._declarationComment());
@@ -1532,7 +1539,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
             currentOffset = line.indexOf(searchString, currentOffset);
             let referenceDetails: RememberedToken | undefined = undefined;
             if (allowLocal && this.semanticFindings.isLocalToken(namePart)) {
-              referenceDetails = this.semanticFindings.getLocalToken(namePart);
+              referenceDetails = this.semanticFindings.getLocalTokenForLine(namePart, lineNbr);
               if (showDebug) {
                 this._logMessage("  --  FOUND local name=[" + namePart + "]");
               }
@@ -1834,7 +1841,6 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
       ptTokenModifiers: declModifiers,
     });
     this._logSPIN("-reportPubPriSig: methodName=[" + methodName + "](" + startNameOffset + ")");
-    this.semanticFindings.clearLocalTokens();
     // -----------------------------------
     //   Parameters
     //
@@ -1864,7 +1870,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
             ptTokenModifiers: ["declaration", "readonly", "local"],
           });
           // remember so we can ID references
-          this.semanticFindings.setLocalToken(paramName, new RememberedToken("parameter", ["readonly", "local"]), lineNbr, this._declarationComment()); // TOKEN SET in _report()
+          this.semanticFindings.setLocalTokenForMethod(methodName, paramName, new RememberedToken("parameter", ["readonly", "local"]), lineNbr, this._declarationComment()); // TOKEN SET in _report()
           currentOffset += paramName.length + 1;
         }
       }
@@ -1906,7 +1912,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
           ptTokenModifiers: ["declaration", "local"],
         });
         // remember so we can ID references
-        this.semanticFindings.setLocalToken(returnValueName, new RememberedToken("returnValue", ["local"]), lineNbr, this._declarationComment()); // TOKEN SET in _report()
+        this.semanticFindings.setLocalTokenForMethod(methodName, returnValueName, new RememberedToken("returnValue", ["local"]), lineNbr, this._declarationComment()); // TOKEN SET in _report()
         currentOffset += returnValueName.length + 1; // +1 for trailing comma
       }
     }
@@ -1954,7 +1960,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
               if (namedIndexPart.substr(0, 1).match(/[a-zA-Z_]/)) {
                 let referenceDetails: RememberedToken | undefined = undefined;
                 if (this.semanticFindings.isLocalToken(namedIndexPart)) {
-                  referenceDetails = this.semanticFindings.getLocalToken(namedIndexPart);
+                  referenceDetails = this.semanticFindings.getLocalTokenForLine(namedIndexPart, lineNbr);
                   this._logSPIN("  --  FOUND local name=[" + namedIndexPart + "]");
                 } else if (this.semanticFindings.isGlobalToken(namedIndexPart)) {
                   referenceDetails = this.semanticFindings.getGlobalToken(namedIndexPart);
@@ -1997,7 +2003,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
               ptTokenModifiers: ["declaration", "local"],
             });
             // remember so we can ID references
-            this.semanticFindings.setLocalToken(localName, new RememberedToken("variable", ["local"]), lineNbr, this._declarationComment()); // TOKEN SET in _report()
+            this.semanticFindings.setLocalTokenForMethod(methodName, localName, new RememberedToken("variable", ["local"]), lineNbr, this._declarationComment()); // TOKEN SET in _report()
           } else {
             // have modifier!
             if (this.parseUtils.isStorageType(localName)) {
@@ -2111,7 +2117,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
                 } else {
                   let referenceDetails: RememberedToken | undefined = undefined;
                   if (this.semanticFindings.isLocalToken(variableNamePart)) {
-                    referenceDetails = this.semanticFindings.getLocalToken(variableNamePart);
+                    referenceDetails = this.semanticFindings.getLocalTokenForLine(variableNamePart, lineNbr);
                     this._logSPIN("  --  FOUND local name=[" + variableNamePart + "]");
                   } else if (this.semanticFindings.isGlobalToken(variableNamePart)) {
                     referenceDetails = this.semanticFindings.getGlobalToken(variableNamePart);
@@ -2172,7 +2178,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
               }
               let referenceDetails: RememberedToken | undefined = undefined;
               if (this.semanticFindings.isLocalToken(cleanedVariableName)) {
-                referenceDetails = this.semanticFindings.getLocalToken(cleanedVariableName);
+                referenceDetails = this.semanticFindings.getLocalTokenForLine(cleanedVariableName, lineNbr);
                 this._logSPIN("  --  FOUND local name=[" + cleanedVariableName + "]");
               } else if (this.semanticFindings.isGlobalToken(cleanedVariableName)) {
                 referenceDetails = this.semanticFindings.getGlobalToken(cleanedVariableName);
@@ -2283,7 +2289,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
           this._logSPIN("  --  SPIN RHS    nameOffset=(" + nameOffset + "), currNonStringOffset=(" + currNonStringOffset + "), currentOffset=(" + currentOffset + ")");
           let referenceDetails: RememberedToken | undefined = undefined;
           if (this.semanticFindings.isLocalToken(namePart)) {
-            referenceDetails = this.semanticFindings.getLocalToken(namePart);
+            referenceDetails = this.semanticFindings.getLocalTokenForLine(namePart, lineNbr);
             this._logSPIN("  --  FOUND local name=[" + namePart + "]");
           } else if (this.semanticFindings.isGlobalToken(namePart)) {
             referenceDetails = this.semanticFindings.getGlobalToken(namePart);
@@ -2507,7 +2513,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
                   referenceDetails = this.semanticFindings.getLocalPasmTokenForMethod(this.currentMethodName, namePart);
                   this._logPASM("  --  FOUND local PASM name=[" + namePart + "]");
                 } else if (this.semanticFindings.isLocalToken(namePart)) {
-                  referenceDetails = this.semanticFindings.getLocalToken(namePart);
+                  referenceDetails = this.semanticFindings.getLocalTokenForLine(namePart, lineNbr);
                   this._logPASM("  --  FOUND local name=[" + namePart + "]");
                 } else if (this.semanticFindings.isGlobalToken(namePart)) {
                   referenceDetails = this.semanticFindings.getGlobalToken(namePart);