+++ spin2/src/spin2.hover.behavior.ts
@@ -11,7 +11,7 @@ import { IPairs, IDefinitionInfo, IDefinitionInput, ExtensionUtils, getSpin2Conf
 
 export class Spin2HoverProvider implements HoverProvider {
   private spinConfig: WorkspaceConfiguration | undefined;
-  private hoverLogEnabled: boolean = false; // WARNING (REMOVE BEFORE FLIGHT)- change to 'false' - disable before commit
+  private hoverLogEnabled: boolean = true; // WARNING (REMOVE BEFORE FLIGHT)- change to 'false' - disable before commit
   private hoverOutputChannel: vscode.OutputChannel | undefined = undefined;
   private symbolsFound: DocumentFindings;
   private parseUtils = new ParseUtils();
@@ -189,7 +189,7 @@ export class Spin2HoverProvider implements HoverProvider {
       }
       if ((bFoundParseToken || bFoundDebugToken) && !builtInFindings.found) {
         bFoundSomething = true;
-        let tokenFindings: ITokenDescription = this.symbolsFound.getTokenWithDescription(input.word);
+        let tokenFindings: ITokenDescription = this.symbolsFound.getTokenWithDescription(input.word, input.position.line + 1);
         if (bFoundDebugToken) {
           tokenFindings = this.symbolsFound.getDebugTokenWithDescription(input.word);
         }
@@ -356,7 +356,7 @@ export class Spin2HoverProvider implements HoverProvider {
           }
           if (titleText && subTitleText) {
             if (builtInFindings.type == eBuiltInType.BIT_CONSTANT && bFoundParseToken) {
-              const tokenFindings = this.symbolsFound.getTokenWithDescription(input.word);
+              const tokenFindings = this.symbolsFound.getTokenWithDescription(input.word, input.position.line + 1);
               if (tokenFindings.found) {
                 const declLine = input.document.lineAt(tokenFindings.declarationLine).text.trim(); // declaration line
                 const nonCommentDecl: string = this.parseUtils.getNonCommentLineRemainder(0, declLine).trim();