+++ spin2/CHANGELOG.md
@@ -20,6 +20,12 @@ Possible next additions:
 - Add new-file templates as Snippets
 - Add additional Snippets as the community identifies them
 
+## [1.9.10] 2023-07-22
+
+General repairs to Hover Help (P1 and P2)
+
+- P1 & P2: Repair hover help - fixed hover display for method parameters, return values, and local variables
+
 ## [1.9.9] 2023-07-17
 
 General repairs to Signature Help (P1 and P2)